% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cI}
\alias{cI}
\title{Friendly Check Box Widget Function}
\usage{
cI(name, start = FALSE)
}
\arguments{
\item{name}{character, the name and label to be used for this widget, must be a proper R variable name}

\item{start}{logical, initial value of the input}
}
\value{
unevaluated expression to be used in generating code for \code{\link{interact}}
}
\description{
Alias for \code{\link{checkboxInput}}(name, name, start)
}
\seealso{
\code{\link{checkboxInput}}, \code{\link{interact}}
}
\keyword{box,}
\keyword{check}
\keyword{checkboxInput,}
\keyword{widget}

