% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dI}
\alias{dI}
\title{Friendly Drop Down Box Widget Function}
\usage{
dI(name, type, start = NULL, ...)
}
\arguments{
\item{name}{character, the name and label to be used for this widget, must be a proper R variable name}

\item{type}{character, the object class for the options, only supports numeric and character}

\item{start}{character or numeric, initial radio button selected}

\item{...}{comma delimited list of possible values for the drop down box}
}
\value{
unevaluated function to be used in generating code for \code{\link{interact}}
}
\description{
Alias for \code{\link{selectInput}}(name, name, ..., selected = start) additional functionality
permits radio buttons for specific numeric inputs
}
\seealso{
\code{\link{radioButtons}}, \code{\link{interact}}
}
\keyword{box,}
\keyword{down}
\keyword{drop}
\keyword{selectInput,}
\keyword{widget}

