% Generated by roxygen2 (4.0.0): do not edit by hand
\name{findWidgets}
\alias{findWidgets}
\title{findWidgets}
\usage{
findWidgets(text)
}
\arguments{
\item{text}{character string of code to be searched}
}
\value{
character vector of each widget function call
}
\description{
Auxillary function for text manipulation. Finds and extracts a complete widget function
call by bracket matching
}

