% Generated by roxygen2 (4.0.0): do not edit by hand
\name{interact}
\alias{interact}
\title{Interact}
\usage{
interact(code, outputType)
}
\arguments{
\item{code}{the code to be made interactive}

\item{outputType}{character specification of output type, "character" or "numeric"}
}
\description{
Specify a code chunk as interactive and generate an applet in a new window
}
\examples{
interact({
plot(1:5 * sI("multiplier", 3, min = 0, max = 5, step = .01), 1:5,
    xlab = tI("xlabel", "Woah"),
    xlim = c(0, 25))
},
outputType = "plot")

interact({
 plot(1:5 * dI("multiplier", "numeric", 3, list(1,2,3)), 1:5,
      xlab = tI("xlabel", "Woah"),
      ylab = as.character(cI("log", FALSE)),
      xlim =c(0,15))
},
outputType = "plot")
}
\keyword{interact,}
\keyword{reactive}
\keyword{shiny,}

