% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matchPull}
\alias{matchPull}
\title{matchPull}
\usage{
matchPull(pattern, text, invert = FALSE, global = FALSE, ...)
}
\arguments{
\item{pattern}{regular expression to be extracted}

\item{text}{character vector to be searched}

\item{invert}{logical, indicating whether to extract matched substring (invert = FALSE) or remainder (default FALSE)}

\item{global}{logical, find multiple matches within a character string (default FALSE)}
}
\value{
character vector of matches or if text has more than one element and global is true a list of vectors of matches
}
\description{
Auxillary function for text manipulation. Finds and extracts a substring specified
by a regular expression. Can be applied arbitrary length vectors
}
\examples{
cat_function()
}
\seealso{
\code{\link{regex}}, \code{\link{regexpr}}, \code{\link{regmatches}}
}
\keyword{regex}
\keyword{substring,}

