% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nI}
\alias{nI}
\title{Friendly Numeric Input Widget Function}
\usage{
nI(name, start = NULL)
}
\arguments{
\item{name}{character, the name and label to be used for this widget, must be a proper R variable name}

\item{start}{numeric, initial value of the input}
}
\value{
unevaluated function to be used in generating code for \code{\link{interact}}
}
\description{
Alias for \code{\link{numericInput}}(name, name, start)
}
\seealso{
\code{\link{numericInput}}, \code{\link{interact}}
}
\keyword{numericInput,}
\keyword{widget}

