% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rI}
\alias{rI}
\title{Friendly Radio Button Widget Function}
\usage{
rI(name, type, start, ...)
}
\arguments{
\item{name}{character, the name and label to be used for this widget, must be a proper R variable name}

\item{type}{character, the object class for the options, only supports numeric and character}

\item{start}{character or numeric, initial radio button selected}

\item{...}{comma delimited list of possible values for the radio buttons}
}
\value{
unevaluated function to be used in generating code for \code{\link{interact}}
}
\description{
Alias for \code{\link{radioButtons}}(name, name, ..., selected = start) additional functionality
permits radio buttons for specific numeric inputs
}
\seealso{
\code{\link{radioButtons}}, \code{\link{interact}}
}
\keyword{buttons,}
\keyword{radio}
\keyword{radioButtons,}
\keyword{widget}

