% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sI}
\alias{sI}
\title{Friendly Slider Widget Function}
\usage{
sI(name, start, min = NULL, max = NULL, step = NULL)
}
\arguments{
\item{name}{character, the name and label to be used for this widget, must be a proper R variable name}

\item{start}{numeric, initial value of the slider}

\item{min}{numeric, minimum value of the slider}

\item{max}{numeric, maximum value of the slider}

\item{step}{numeric, smallest step slider can move (optional)}
}
\value{
unevaluated function to be used in generating code for \code{\link{interact}}
}
\description{
Alias for \code{\link{sliderInput}}(name, name, min, max, step)
}
\seealso{
\code{\link{sliderInput}}, \code{\link{interact}}
}
\keyword{slider,}
\keyword{sliderInput,}
\keyword{widget}

