% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tI}
\alias{tI}
\title{Friendly Text Input Widget Function}
\usage{
tI(name, start = "")
}
\arguments{
\item{name}{character, the name and label to be used for this widget, must be a proper R variable name}

\item{start}{character, initial value of the input}
}
\value{
unevaluated expression to be used in generating code for \code{\link{interact}}
}
\description{
Alias for \code{\link{textInput}}(name, name, start)
}
\seealso{
\code{\link{textInput}}, \code{\link{interact}}
}
\keyword{textInput,}
\keyword{widget}

